LazyTensor Cleanup - Move graph compiler/execution part out
Summary: Move the following public methods and relevant private methods into
LazyGraphExecutor
- GetLiveTensors
- SyncLiveTensorsGraph
- SyncTensorsGraph
- MarkStep
- WaitDeviceOps
- GetTensors
